The list of users of crtc_hw.h will continue to grow so we place it in the main header. Also the index for the registers is only u8 not u32

+++ drivers/gpu/drm/via/crtc_hw.h
@@ -37,7 +37,7 @@
 #define GRA_I           0x0e
 #define GRA_D           0x0f
 
-static inline void gra_outb(void __iomem *regs, u32 idx, u8 val)
+static inline void gra_outb(void __iomem *regs, u8 idx, u8 val)
 {
 	iowrite8(idx, regs + GRA_I);
 	wmb();
@@ -45,7 +45,7 @@ static inline void gra_outb(void __iomem *regs, u32 idx, u8 val)
 	wmb();
 }
 
-static inline void seq_outb(void __iomem *regs, u32 idx, u8 val)
+static inline void seq_outb(void __iomem *regs, u8 idx, u8 val)
 {
 	iowrite8(idx, regs + SEQ_I);
 	wmb();
@@ -53,14 +53,14 @@ static inline void seq_outb(void __iomem *regs, u32 idx, u8 val)
 	wmb();
 }
 
-static inline u8 seq_inb(void __iomem *regs, u32 idx)
+static inline u8 seq_inb(void __iomem *regs, u8 idx)
 {
 	iowrite8(idx, regs + SEQ_I);
 	mb();
 	return ioread8(regs + SEQ_D);
 }
 
-static inline void crtc_outb(void __iomem *regs, u32 idx, u8 val)
+static inline void crtc_outb(void __iomem *regs, u8 idx, u8 val)
 {
 	iowrite8(idx, regs + CRT_I);
 	wmb();
@@ -68,14 +68,14 @@ static inline void crtc_outb(void __iomem *regs, u32 idx, u8 val)
 	wmb();
 }
 
-static inline u8 crtc_inb(void __iomem *regs, u32 idx)
+static inline u8 crtc_inb(void __iomem *regs, u8 idx)
 {
 	iowrite8(idx, regs + CRT_I);
 	mb();
 	return ioread8(regs + CRT_D);
 }
 
-static inline void att_outb(void __iomem *regs, u32 idx, u8 val)
+static inline void att_outb(void __iomem *regs, u8 idx, u8 val)
 {
 	unsigned char tmp;
